Some validations added for the reservation making, and sending email when reservation is deleted

+++ src/main/java/com/airbnb/controller/reservation/ReservationController.java
@@ -39,12 +39,12 @@ public class ReservationController {
 	private PlaceDAO placeDAO;
 
 	@RequestMapping(value = "/reservation/{id}", method = RequestMethod.GET)
-	public String makeReservation(Model model,@PathVariable("id") int id, HttpServletRequest request) {
+	public String makeReservation(Model model, @PathVariable("id") int id, HttpServletRequest request) {
 		try {
 			HttpSession session = request.getSession();
 			User user = (User) session.getAttribute("user");
 			if (user == null) {
-				return "redirect: ./logout";
+				return "redirect:/logout";
 			}
 			PlaceDTO view = placeDAO.getDtoById(id);
 			session.setAttribute("place", view);
@@ -52,6 +52,7 @@ public class ReservationController {
 			Reservation reservation = new Reservation();
 			session.setAttribute("reservation", reservation);
 			session.setAttribute("wrongDates", false);
+			session.setAttribute("sameUser", false);
 			return "reservation";
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -61,32 +62,56 @@ public class ReservationController {
 	}
 
 	@RequestMapping(value = "/reservation", method = RequestMethod.POST)
-	public String makeReservation(Model model, @RequestParam Date startDate, @RequestParam Date endDate, HttpServletRequest request) {
+	public String makeReservation(Model model, @RequestParam Date startDate, @RequestParam Date endDate,
+			HttpServletRequest request) {
 		try {
-		HttpSession session = request.getSession();
-		User user = (User) session.getAttribute("user");
-		if (user == null) {
-			return "index";
-		}
-		LocalDate start = startDate.toLocalDate();
-		LocalDate end = endDate.toLocalDate();
-		if ((end.isBefore(start)) ||( start.isBefore(LocalDate.now()))) {
-			session.setAttribute("wrongDates", true);
-			return "reservation";
-		}
-		
-		PlaceDTO place= (PlaceDTO)session.getAttribute("place");
-		reservationDAO.makeReservation(new Reservation(0,start,end,place.getId(),user.getId(),0,LocalDate.now()));
+			HttpSession session = request.getSession();
+			User user = (User) session.getAttribute("user");
+			if (user == null) {
+				return "index";
+			}
+			LocalDate start = startDate.toLocalDate();
+			LocalDate end = endDate.toLocalDate();
+			if ((end.isBefore(start)) || (start.isBefore(LocalDate.now()))) {
+				session.setAttribute("wrongDates", true);
+				return "reservation";
+			}
+			PlaceDTO place = (PlaceDTO) session.getAttribute("place");
+			if (user.getId() == place.getOwnerId()) {
+				session.setAttribute("sameUser", true);
+				return "reservation";
+			}
 
-		short days=(short) start.until(end, ChronoUnit.DAYS);
-		double fullPrice=place.getPrice()*days;
-		
-		String userContent=String.format("You've successfully made reservation for: %s, place type - %s, Address: Country - %s, City - %s, Street - %s, StreetNumber - %d , for %d days and it will cost you %.2f. The owner of the place have 7 days from now to reject your reservation! You can check in your \"Visited places\" section, if he rejects your reservation the place will no longer be there. If the place is there after the 7 days period that means your reservation is completed!", place.getName(),place.getPlaceTypeName(),place.getCountry(),place.getCity(),place.getStreet(),place.getStreetNumber(),days,fullPrice);
-		MailSender.sendEmail(user.getEmail(), userContent);
-		String ownerContent=String.format("%s %s with e-mail: %s, has made a reservation for: %s, place type - %s, Address: Country - %s, City - %s, Street - %s, StreetNumber - %d , for %d days and it will cost %.2f. If you want to reject the reservation you have to log in our site and go to \"Awaiting reservations\" section and then click on the \"Reject  reservation\" button on the reservation! You have 7 days to reject it or it will be automaticly completed", user.getFirstName(),user.getLastName(),user.getEmail(),place.getName(),place.getPlaceTypeName(),place.getCountry(),place.getCity(),place.getStreet(),place.getStreetNumber(),days,fullPrice);
-		MailSender.sendEmail(userDAO.userFromId(place.getOwnerId()).getEmail(), ownerContent);
-		
-		return "redirect: ./search";
+			int reservationId = reservationDAO
+					.makeReservation(new Reservation(0, start, end, place.getId(), user.getId(), 0, LocalDate.now()));
+			Thread t = new Thread(() -> {
+				while (true) {
+					try {
+						Thread.sleep(24 * 60 * 60 * 1000);
+						if (end.isBefore(LocalDate.now())) {
+							reservationDAO.deleteReservation(reservationId);
+						}
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
+				}
+			});
+			t.start();
+			short days = (short) start.until(end, ChronoUnit.DAYS);
+			double fullPrice = place.getPrice() * days;
+
+			String userContent = String.format(
+					"You've successfully made reservation for: %s, place type - %s, Address: Country - %s, City - %s, Street - %s, StreetNumber - %d , for %d days and it will cost you %.2f. The owner of the place have 7 days from now to reject your reservation! You can check in your \"Visited places\" section, if he rejects your reservation the place will no longer be there. If the place is there after the 7 days period that means your reservation is completed!",
+					place.getName(), place.getPlaceTypeName(), place.getCountry(), place.getCity(), place.getStreet(),
+					place.getStreetNumber(), days, fullPrice);
+			MailSender.sendEmail(user.getEmail(), userContent);
+			String ownerContent = String.format(
+					"%s %s with e-mail: %s, has made a reservation for: %s, place type - %s, Address: Country - %s, City - %s, Street - %s, StreetNumber - %d , for %d days and it will cost %.2f. If you want to reject the reservation you have to log in our site and go to \"Awaiting reservations\" section and then click on the \"Reject  reservation\" button on the reservation! You have 7 days to reject it or it will be automaticly completed",
+					user.getFirstName(), user.getLastName(), user.getEmail(), place.getName(), place.getPlaceTypeName(),
+					place.getCountry(), place.getCity(), place.getStreet(), place.getStreetNumber(), days, fullPrice);
+			MailSender.sendEmail(userDAO.userFromId(place.getOwnerId()).getEmail(), ownerContent);
+
+			return "redirect: ./search";
 
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -94,7 +119,7 @@ public class ReservationController {
 			return "error";
 		}
 	}
-	
+
 	@RequestMapping(value = "/reservations", method = RequestMethod.GET)
 	public String getReservations(Model model, HttpSession session) {
 		try {
@@ -102,15 +127,19 @@ public class ReservationController {
 			if (user == null) {
 				return "redirect: ./logout";
 			}
-			Map <Reservation,PlaceDTO>reservations=new TreeMap<Reservation,PlaceDTO>((r1,r2)->{return r2.getId()-r1.getId();});
-			List<Reservation> reservationsList=reservationDAO.getAllReservationsByHostID(user.getId());
-
+			Map<Reservation, PlaceDTO> reservations = new TreeMap<Reservation, PlaceDTO>((r1, r2) -> {
+				return r2.getId() - r1.getId();
+			});
+			List<Reservation> reservationsList = reservationDAO.getAllReservationsByHostID(user.getId());
+			List<User> guests=new ArrayList<User>();
 			for (Reservation reservation : reservationsList) {
-				PlaceDTO place=placeDAO.getDtoById(reservation.getPlaceId());
+				guests.add(userDAO.userFromId(reservation.getUserId()));
+				PlaceDTO place = placeDAO.getDtoById(reservation.getPlaceId());
 				reservations.put(reservation, place);
 			}
 
 			model.addAttribute("reservations", reservations);
+			model.addAttribute("guests",guests);
 			return "reservations";
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -118,16 +147,28 @@ public class ReservationController {
 			return "error";
 		}
 	}
-	
+
 	@RequestMapping(value = "/reservations/{id}", method = RequestMethod.GET)
-	public String deleteReservations(Model model,@PathVariable("id") int id, HttpSession session) {
+	public String deleteReservations(Model model, @PathVariable("id") int id, HttpSession session) {
 		try {
 			User user = (User) session.getAttribute("user");
 			if (user == null) {
 				return "redirect: ./logout";
 			}
+			Reservation reservation = reservationDAO.reservationFromId(id);
+			LocalDate sevenDaysPeriod = reservation.getReservationDate().plusDays(7);
+			if (sevenDaysPeriod.isBefore(LocalDate.now())) {
+				return "redirect:/reservations";
+			}
 			reservationDAO.deleteReservation(id);
-//			TODO send email to the guest
+			short days = (short) reservation.getStartDate().until(reservation.getEndDate(), ChronoUnit.DAYS);
+			PlaceDTO place = placeDAO.getDtoById(reservation.getPlaceId());
+			String userContent = String.format(
+					"REJCETED! Your reservation for: %s, place type - %s, Address: Country - %s, City - %s, Street - %s, StreetNumber - %d , for %d days has been rejected by the owner of the place.",
+					place.getName(), place.getPlaceTypeName(), place.getCountry(), place.getCity(), place.getStreet(),
+					place.getStreetNumber(), days);
+			User u = userDAO.userFromId(reservation.getUserId());
+			MailSender.sendEmail(u.getEmail(), userContent);
 			return "redirect:/reservations";
 		} catch (Exception e) {
 			e.printStackTrace();
